+++ src/nextjs/server/invalidateCache.ts
@@ -5,6 +5,8 @@ import { cookies } from "next/headers";
 export async function invalidateCache() {
   // Dummy cookie, just to set the header which will invalidate
   // the client Router Cache.
-  cookies().delete(`__convexAuthCookieForRouterCacheInvalidation${Date.now()}`);
+  (await cookies()).delete(
+    `__convexAuthCookieForRouterCacheInvalidation${Date.now()}`,
+  );
   return null;
 }